requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aioredis==1.3.1
asgiref==3.3.4; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
autobahn==21.3.1; python_version >= '3.7'
automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
channels-redis==3.2.0
channels==3.0.3
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
constantly==15.1.0
cryptography==3.4.7; python_version >= '3.6'
daphne==3.0.2; python_version >= '3.6'
django-cors-headers==3.7.0
django-filter==2.4.0
django==3.2.4
djangorestframework-simplejwt==4.7.1
djangorestframework==3.12.4
environs==9.3.2
hiredis==2.0.0; python_version >= '3.6'
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
markdown==3.3.4
marshmallow==3.12.1; python_version >= '3.5'
msgpack==1.0.2
psycopg2==2.9.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.1.0; python_version >= '3.6'
pyopenssl==20.0.1
python-dotenv==0.18.0
pytz==2021.1
pyyaml==5.4.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
twisted[tls]==21.2.0; python_full_version >= '3.5.4'
txaio==21.2.1; python_version >= '3.6'
uritemplate==3.0.1
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
